requirements.txt
einops==0.4.1
flags==0.0.1.2
h5py==2.10.0
imageio==2.5.0
matplotlib==3.0.3
monai==0.8.1
nibabel==3.0.2
numpy-stl==3.0.1
open3d-python==0.7.0.0
opencv-python==4.2.0.34
openmesh==1.1.3
pandas==0.25.3
Pillow==6.1.0
plyfile==0.8
protobuf==3.9.2
pydicom==1.3.0
PyInstaller==3.5
pytools==2019.1.1
PyYAML==5.1.2
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tornado==6.1
trimesh==3.9.29
urllib3==1.26.2
vedo==2023.4.6
vtk==9.0.1